requirements.txt
requests
numpy
pandas
opencv-python-headless
transformers==4.25.1
timm==0.5.4
albumentations>=1.0.0
pypdf
# lightning>=2.0.0,<2022
sconf>=0.2.3
nltk
fitz
jsonlines 
pathlib
Pillow
PyMuPDF
Levenshtein
matplotlib
orjson
Logger
wandb
pytorch-lightning
# -e .

